# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document Search API",
        "description": "A powerful document search API engine using TNT Search for fast full-text search capabilities. This API supports multi-tenant workspaces, document upload, text extraction, and advanced search features.",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List all projects in workspace",
                "description": "Get a paginated list of all projects in the workspace",
                "operationId": "getProjects",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Project"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Create a new project in the workspace",
                "operationId": "storeProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Project name",
                                        "type": "string",
                                        "example": "Marketing Materials"
                                    },
                                    "description": {
                                        "description": "Project description",
                                        "type": "string",
                                        "example": "All marketing related documents"
                                    },
                                    "color": {
                                        "description": "Hex color for UI",
                                        "type": "string",
                                        "example": "#3B82F6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Project"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a specific project",
                "description": "Retrieve details of a specific project by ID",
                "operationId": "getProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Project"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
        "/api/projects/{id}/statistics": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project statistics",
                "description": "Get detailed statistics for a specific project",
                "operationId": "getProjectStatistics",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "properties": {
                                                        "total_documents": {
                                                            "type": "integer"
                                                        },
                                                        "indexed_documents": {
                                                            "type": "integer"
                                                        },
                                                        "unindexed_documents": {
                                                            "type": "integer"
                                                        },
                                                        "total_size": {
                                                            "type": "integer"
                                                        },
                                                        "formatted_total_size": {
                                                            "type": "string"
                                                        },
                                                        "recent_uploads": {
                                                            "type": "integer"
                                                        },
                                                        "file_types": {
                                                            "type": "object"
                                                        },
                                                        "last_indexed_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "WorkspaceKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
...
            "Project": {
                "title": "Project",
                "description": "Project model",
                "properties": {
                    "id": {
                        "description": "Project ID",
                        "type": "integer",
                        "example": 1
                    },
                    "workspace_id": {
                        "description": "Workspace ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Project name",
                        "type": "string",
                        "example": "My Project"
                    },
                    "slug": {
                        "description": "Project slug",
                        "type": "string",
                        "example": "my-project"
                    },
                    "description": {
                        "description": "Project description",
                        "type": "string",
                        "example": "A sample project for document management",
                        "nullable": true
                    },
                    "color": {
                        "description": "Project color (hex code)",
                        "type": "string",
                        "example": "#007bff",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Whether the project is active",
                        "type": "boolean",
                        "example": true
                    },
                    "settings": {
                        "description": "Project settings (JSON object)",
                        "type": "object",
                        "example": {
                            "indexing_enabled": true,
                            "auto_extract": false
                        },
                        "nullable": true
                    },
                    "last_indexed_at": {
                        "description": "Last indexing timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T10:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "WorkspaceKey": {
                "type": "apiKey",
                "description": "Workspace API key for authentication",
                "name": "X-Workspace-Key",
                "in": "header"
            }
        }
    },
...
}